requirements.txt
matrix-nio
pymongo
